+++ NEWS.md
@@ -46,6 +46,8 @@ did not match the unit (\#823).
 - When creating a scenario, the name of the scenario is set as the name of the simulation.
 This way, when saving the simulation to PKML and loading in MoBi, the loaded simulation
 will have the updated name.
+- Fixed a bug in `createPlotsFromExcel()` when subtitle of PlotConfiguration was
+not applied (\#845).
 
 # esqlabsR 5.4.0